Rename algorithms slicer file names to match function name
