+++ packages/elasticsearch-asset-apis/src/elasticsearch-reader-api/algorithms/idSlicer.ts
