+++ packages/elasticsearch-asset-apis/src/elasticsearch-reader-api/algorithms/index.ts
@@ -1,5 +1,5 @@
 export * from './date-helpers';
-export * from './date-slicer';
+export * from './dateSlicer';
 export * from './id-helpers';
-export * from './id-slicer';
+export * from './idSlicer';
 export * from './utils';
